+++ tests/Options/Special/DateTimeOptionTest.php
@@ -47,7 +47,9 @@ class DateTimeOptionTest extends \WP_UnitTestCase
                 try {
                     $this->stub->flush();
                 } catch (\Exception $exception) {
-                    $this->assertTrue(is_a($exception, $expected));
+                    $this->assertInstanceOf($expected, $exception);
+                } finally {
+                    $this->assertInstanceOf($expected, $exception);
                 }
             }
         }
@@ -61,6 +63,6 @@ class DateTimeOptionTest extends \WP_UnitTestCase
     public function testNull()
     {
         $this->stub->set(null);
-        $this->assertEquals('', $this->stub->get());
+        $this->assertNull($this->stub->get());
     }
 }