+++ tests/PostMeta/Special/BoolPostMetaTest.php
@@ -65,7 +65,7 @@ class BoolPostMetaTest extends \WP_UnitTestCase
             }
         } else {
             $this->stub->flush();
-            $this->assertEquals($expected, $this->stub->get());
+            $this->assertSame($expected, $this->stub->get());
         }
     }
 
@@ -85,9 +85,9 @@ class BoolPostMetaTest extends \WP_UnitTestCase
         $this->stub->set($value);
 
         if (class_exists($expected)) {
-            $this->assertEquals($value, $this->stub->get());
+            $this->assertSame($value, $this->stub->get());
         } else {
-            $this->assertEquals($expected, $this->stub->get());
+            $this->assertSame($expected, $this->stub->get());
         }
     }
 
@@ -98,6 +98,6 @@ class BoolPostMetaTest extends \WP_UnitTestCase
 
     public function testDefaultValue()
     {
-        $this->assertEquals(true, $this->stub->get());
+        $this->assertSame(true, $this->stub->get());
     }
 }